Introduce fmt() function
