+++ src/misc/utils.ts
@@ -1,5 +1,4 @@
-import { pp, PPIsh } from "..";
-import { ASTNode } from "../ast/ast_node";
+import { fmt, PPIsh } from "..";
 
 export function forAll<T>(iterable: Iterable<T>, cb: (v: T) => boolean): boolean {
     for (const el of iterable) {
@@ -30,28 +29,5 @@ export function assert(
         return;
     }
 
-    if (details.length) {
-        const nodes: ASTNode[] = [];
-
-        for (let i = 0; i < details.length; i++) {
-            const detail = details[i];
-            const part = pp(detail);
-
-            if (detail instanceof ASTNode) {
-                nodes.push(detail);
-            }
-
-            message = message.replace(new RegExp("\\{" + i + "\\}", "g"), part);
-        }
-
-        if (nodes.length) {
-            if (!message.endsWith(".")) {
-                message += ".";
-            }
-
-            message += "\n\n" + nodes.map((node) => node.print()).join("\n");
-        }
-    }
-
-    throw new Error(message);
+    throw new Error(fmt(message, ...details));
 }
